commit combos provincias completo

+++ src/java/model/Barrio.java
@@ -12,19 +12,19 @@ package model;
 public class Barrio {
 
     //Definicion de variables
-    int COD_PROVINCIA;
-    int COD_CANTON;
-    int COD_DISTRITO;
-    int COD_BARRIO;
+    float COD_PROVINCIA;
+    float COD_CANTON;
+    float COD_DISTRITO;
+    float COD_BARRIO;
     String DSC_BARRIO;
-    int LOG_ACTIVO;
+    float LOG_ACTIVO;
     
     //Metodo constructor
 
     public Barrio() {
     }
 
-    public Barrio(int COD_PROVINCIA, int COD_CANTON, int COD_DISTRITO, int COD_BARRIO, String DSC_BARRIO, int LOG_ACTIVO) {
+    public Barrio(float COD_PROVINCIA, float COD_CANTON, float COD_DISTRITO, float COD_BARRIO, String DSC_BARRIO, float LOG_ACTIVO) {
         this.COD_PROVINCIA = COD_PROVINCIA;
         this.COD_CANTON = COD_CANTON;
         this.COD_DISTRITO = COD_DISTRITO;
@@ -35,35 +35,35 @@ public class Barrio {
     
     //Metodos SET y GET
 
-    public int getCOD_PROVINCIA() {
+    public float getCOD_PROVINCIA() {
         return COD_PROVINCIA;
     }
 
-    public void setCOD_PROVINCIA(int COD_PROVINCIA) {
+    public void setCOD_PROVINCIA(float COD_PROVINCIA) {
         this.COD_PROVINCIA = COD_PROVINCIA;
     }
 
-    public int getCOD_CANTON() {
+    public float getCOD_CANTON() {
         return COD_CANTON;
     }
 
-    public void setCOD_CANTON(int COD_CANTON) {
+    public void setCOD_CANTON(float COD_CANTON) {
         this.COD_CANTON = COD_CANTON;
     }
 
-    public int getCOD_DISTRITO() {
+    public float getCOD_DISTRITO() {
         return COD_DISTRITO;
     }
 
-    public void setCOD_DISTRITO(int COD_DISTRITO) {
+    public void setCOD_DISTRITO(float COD_DISTRITO) {
         this.COD_DISTRITO = COD_DISTRITO;
     }
 
-    public int getCOD_BARRIO() {
+    public float getCOD_BARRIO() {
         return COD_BARRIO;
     }
 
-    public void setCOD_BARRIO(int COD_BARRIO) {
+    public void setCOD_BARRIO(float COD_BARRIO) {
         this.COD_BARRIO = COD_BARRIO;
     }
 
@@ -75,12 +75,13 @@ public class Barrio {
         this.DSC_BARRIO = DSC_BARRIO;
     }
 
-    public int getLOG_ACTIVO() {
+    public float getLOG_ACTIVO() {
         return LOG_ACTIVO;
     }
 
-    public void setLOG_ACTIVO(int LOG_ACTIVO) {
+    public void setLOG_ACTIVO(float LOG_ACTIVO) {
         this.LOG_ACTIVO = LOG_ACTIVO;
     }
+
     
 }
